Largest complex: sort snapshots by time
- When plotting the evolution of the largest complex over a set of
snapshots, sort the data by the timestamp of the snapshot. When using
the `-mixture` option of KaSim, the T0 time is advanced to the
snapshot's, but the event number is not. Thus event number, and by
extension snapshot name, is not enough to sort snapshots. Re-using code
from the multithreaded snapshot reader.